requirements.txt
azure-functions
pydantic
markdown
azure-identity
azure.ai.language.questionanswering
aiohttp>=3.0
azure.storage.blob